Create rockpaperscissor.go
It is a simple rock paper or scissor game, where you can choose how many turns you want,

It has one bug that I dont now yet how to fix and that is if you enter text instead of numbers when it asks you for how many turns you want. It use that string for some reason. I have made it set to default one game, but it is still a little buggy.